Add representation for meta properties
ECMAScript gets new syntax thing call "meta properties", which is already used as `new.target` and, according to meeting notes, can be used in future in other places as well (i.e. `yield.input`).

For those cases, we need new node that (up to discussion) derives from `MemberExpression` but is restricted to having `Identifier` in both `object` (as it's actually a keyword used as `Identifier`) and `property` and can't be `computed` (i.e. `new['target']` is syntax error).